+++ packages/application-shell/src/constants.ts
@@ -1,5 +1,7 @@
-export const HEIGHTS = {
+export const DIMENSIONS = {
   header: '43px',
+  navMenu: '64px',
+  navMenuExpanded: '200px',
 } as const;
 
 export const SUPPORTED_HEADERS = {