+++ visual-testing-app/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @commercetools-local/visual-testing-app
 
+## 19.2.3
+
+### Patch Changes
+
+- [#2546](https://github.com/commercetools/merchant-center-application-kit/pull/2546) [`dc76e5a9`](https://github.com/commercetools/merchant-center-application-kit/commit/dc76e5a9a7f875dadf2ed5a11c48a1ddff7b431c) Thanks [@renovate](https://github.com/apps/renovate)! - Upgrade dependencies
+
+- Updated dependencies [[`54018ddf`](https://github.com/commercetools/merchant-center-application-kit/commit/54018ddf8a953ff9409c82ab3fe3d63feb158500), [`dc76e5a9`](https://github.com/commercetools/merchant-center-application-kit/commit/dc76e5a9a7f875dadf2ed5a11c48a1ddff7b431c)]:
+  - @commercetools-frontend/application-components@21.3.4
+  - @commercetools-frontend/constants@21.3.4
+  - @commercetools-frontend/react-notifications@21.3.4
+
 ## 19.2.2
 
 ### Patch Changes